requirements.txt
langchain>=0.1.13
langchain-openai>=0.0.5
openai>=1.14.3
faiss-cpu
python-dotenv
chromadb>=0.4.24
